Simplifying the web setup script & renaming `setup-rn-audio-api-web`
